pyproject.toml
[tool.poetry]
name = "mistapp"
version = "0.1.0"
description = "No"
authors = ["Isai Gordeev <isai.gordeev.2022@polytechnique.org>"]
license = "No License"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.8"
uvicorn = "^0.34.0"
requests = "^2.32.3"
httpx = "^0.28.1"
mistralai = "^1.5.0"
python-dotenv = "^1.0.1"
pytest = "^8.3.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
